requirements.txt
discord.py
flask
apscheduler
python-dotenv
